Restructure resolvers/definitions
