+++ services/app/app/gql/queries/config.graphql
@@ -2,7 +2,6 @@ query CompanyUpdateConfig {
   companyUpdateConfig {
     contactUrl
     logoUrl
-    leadershipSectionIds
     leadershipSectionMax
   }
 }
