+++ services/graphql/src/definitions/config.js
@@ -1,0 +1,15 @@
+const { gql } = require('apollo-server');
+
+module.exports = gql`
+
+  extend type Query {
+    companyUpdateConfig: CompanyUpdateConfiguration!
+  }
+
+  type CompanyUpdateConfiguration {
+    logoUrl: String
+    contactUrl: String
+    leadershipSectionMax: Int!
+  }
+
+`;
