+++ services/graphql/src/resolvers/config.js
@@ -1,0 +1,11 @@
+const { LOGO_URL, CONTACT_URL, LEADERSHIP_SECTION_MAX } = require('../env');
+
+module.exports = {
+  Query: {
+    companyUpdateConfig: () => ({
+      contactUrl: CONTACT_URL,
+      logoUrl: LOGO_URL,
+      leadershipSectionMax: LEADERSHIP_SECTION_MAX,
+    }),
+  },
+};
