+++ services/graphql/src/resolvers/index.js
@@ -1,79 +1,38 @@
+const deepAssign = require('deep-assign');
+// const GraphQLJSON = require('graphql-type-json');
 const { DateType } = require('@limit0/graphql-custom-types');
-const uuid = require('uuid/v4');
-const { notify, thank } = require('../mailer');
-const env = require('../env');
-const s3Client = require('../s3-client');
-const {
-  retrieve,
-  insert,
-  complete,
-  submissions,
-  submissionCount,
-} = require('../mongodb');
 
-const {
-  LOGO_URL,
-  CONTACT_URL,
-  LEADERSHIP_SECTION_IDS,
-  LEADERSHIP_SECTION_MAX,
-  TENANT_KEY,
-  AWS_S3_BUCKET,
-} = env;
+const company = require('./company');
+const config = require('./config');
+const leadership = require('./leadership');
+const list = require('./list');
 
-const config = {
-  contactUrl: CONTACT_URL,
-  logoUrl: LOGO_URL,
-  leadershipSectionIds: LEADERSHIP_SECTION_IDS,
-  leadershipSectionMax: LEADERSHIP_SECTION_MAX,
-};
+module.exports = deepAssign(
+  company,
+  config,
+  leadership,
+  list,
+  {
+    Date: DateType,
+    // JSON: GraphQLJSON, // @todo is this needed?
 
-module.exports = {
-  Date: DateType,
-
-  CompanyUpdateSubmission: {
-    id: ({ _id }) => _id.toString(),
-    submitted: ({ _id }) => _id.getTimestamp(),
-    payload: ({ payload }) => JSON.stringify(payload),
-    companyName: ({ payload: { name } }) => name,
-  },
-
-  Query: {
-    companyUpdateConfig: () => config,
-    companyUpdateSubmission: (_, { id }) => retrieve(id),
-    companyUpdateSubmissions: (_, { input }) => submissions(input.all),
-    companyUpdateSubmissionCount: () => submissionCount(),
-  },
-  Mutation: {
-    companyUpdateSubmit: async (_, { input }) => {
-      const reviewed = false;
-      const res = await insert({ reviewed, ...input });
-      if (!res.result.ok) throw new Error('Unable to save your changes, please try again.');
-      const submission = await retrieve(res.insertedId);
-      await notify(submission);
-      await thank(submission);
-      return submission;
+    /**
+     *
+     */
+    Query: {
+      /**
+       *
+       */
+      ping: () => 'pong',
     },
-    companyUpdateComplete: async (_, { id }) => {
-      await complete(id);
-      return 'ok';
-    },
-    companyUpdateDiscard: async (_, { id }) => {
-      await complete(id);
-      return 'ok';
-    },
-    companyUpdateSingleUpload: async (_, { file }) => {
-      const { createReadStream, filename, mimetype } = await file;
-      const Bucket = AWS_S3_BUCKET;
-      const Key = `${TENANT_KEY}/${uuid()}/${filename}`;
-
-      const response = await s3Client.upload({
-        Bucket,
-        Key,
-        Body: createReadStream(),
-        ACL: 'public-read',
-        ContentType: mimetype,
-      }).promise();
-      return response.Location;
+    /**
+     *
+     */
+    Mutation: {
+      /**
+       *
+       */
+      ping: () => 'pong',
     },
   },
-};
+);
