+++ services/graphql/src/resolvers/leadership.js
@@ -1,0 +1,11 @@
+module.exports = {
+  /**
+   *
+   */
+  Query: {
+    /**
+     *
+     */
+    companyUpdateLeadership: () => ({}),
+  },
+};
