+++ services/graphql/src/resolvers/list.js
@@ -1,0 +1,18 @@
+const { submissions, submissionCount } = require('../mongodb');
+
+module.exports = {
+  /**
+   *
+   */
+  Query: {
+    /**
+     *
+     */
+    companyUpdateSubmissions: (_, { input }) => submissions(input.all),
+
+    /**
+     *
+     */
+    companyUpdateSubmissionCount: () => submissionCount(),
+  },
+};
